+++ c#/ccxt/base/Throttler.cs
@@ -10,7 +10,7 @@ public class Throttler
 {
 
     private dict config = new dict();
-    private Queue<(Task, float)> queue = new Queue<(Task, float)>();
+    private Queue<(Task, double)> queue = new Queue<(Task, double)>();
 
     private bool running = false;
 
@@ -43,7 +43,7 @@ public class Throttler
             var first = this.queue.Peek();
             var task = first.Item1;
             var cost = first.Item2;
-            var floatTokens = float.Parse(this.config["tokens"].ToString(), CultureInfo.InvariantCulture);
+            var floatTokens = double.Parse(this.config["tokens"].ToString(), CultureInfo.InvariantCulture);
             if (floatTokens >= 0)
             {
                 this.config["tokens"] = floatTokens - cost;
@@ -61,7 +61,7 @@ public class Throttler
                 var current = milliseconds();
                 var elapsed = current - lastTimestamp;
                 lastTimestamp = current;
-                var tokens = (float)this.config["tokens"] + ((float)this.config["refillRate"] * elapsed);
+                var tokens = (double)this.config["tokens"] + ((double)this.config["refillRate"] * elapsed);
                 this.config["tokens"] = Math.Min(tokens, (int)this.config["capacity"]);
             }
         }
@@ -70,13 +70,13 @@ public class Throttler
 
     public async Task<Task> throttle(object cost2)
     {
-        var cost = (cost2 != null) ? float.Parse(cost2.ToString(), CultureInfo.InvariantCulture) : this.config["cost"];
+        var cost = (cost2 != null) ? double.Parse(cost2.ToString(), CultureInfo.InvariantCulture) : this.config["cost"];
         if (this.queue.Count > (int)this.config["maxCapacity"])
         {
             throw new Exception("throttle queue is over maxCapacity (" + this.config["maxCapacity"].ToString() + "), see https://github.com/ccxt/ccxt/issues/11645#issuecomment-1195695526");
         }
         var t = new Task(() => { });
-        this.queue.Enqueue((t, (float)cost));
+        this.queue.Enqueue((t, (double)cost));
         if (!this.running)
         {
             this.running = true;
